refactor: marathon, move SD struct from config into marathon discovery package
